mae needed changes in circlesubs

+++ src/main/java/cmu/curantis/dao/CircleSubsDAO.java
@@ -15,7 +15,7 @@ public class CircleSubsDAO {
 	/*
 	 * Returns the new circle id for this circle. Returns -1 on error.
 	 */
-	public int addCircleSubs(Session session, CircleSubsBean bean){
+	public long addCircleSubs(Session session, CircleSubsBean bean){
 		session.save(bean);
 		List<CircleSubsBean> result = session.createQuery("from CircleSubsBean ORDER BY circle_id DESC").setMaxResults(1).list();
 		if (result == null || result.size() == 0) {
